feat(ecs): Added ECS

+++ src/main/scala/github/dbousamra/ecs/Systems.scala
@@ -4,5 +4,5 @@ trait System {
 
   var engine: Option[Engine] = None
 
-  def process(es: List[Entity], delta: Int): Unit
+  def process(es: List[Entity], delta: Float): Unit
 }
